src/v1/services/video_preprocessor/requirements.txt
aio-pika==6.7.1
aiormq==3.3.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
facenet-pytorch==2.5.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.0
opencv-python==4.5.1.48
pamqp==2.3.0
pika==1.1.0
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
scipy==1.6.0
six==1.15.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.3
watchdog==1.0.2
yarl==1.6.3
